ai_agent/requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio>=3.4.3
beautifulsoup4>=4.9.3
feedparser>=6.0.0
python-dotenv>=0.19.0
requests>=2.26.0
spacy>=3.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

# AI and NLP
protobuf>=4.23.0
torch>=1.9.0
tenacity-9.1.2
torchvision==0.14.1
torchaudio==0.13.1
transformers>=4.11.0
sentence-transformers>=2.1.0
huggingface-hub==0.13.4
sentencepiece==0.1.99
accelerate==0.17.1
langchain>=0.0.200

# Text Processing
textblob>=0.15.3
nltk>=3.6.0

# Database 
sqlalchemy==2.0.19
alembic==1.11.1
aiosqlite>=0.19.0

# Core utilities
numpy>=1.19.5
pandas==2.0.3
regex==2023.6.3
pytz==2023.3
langdetect==1.0.9
python-slugify>=5.0.2
tqdm==4.65.0
pytrends>=0.9.0
psutil>=5.8.0
PyYAML>=5.4.1

# Web & Scraping
trafilatura>=0.8.2
newspaper3k>=0.2.8
brotli>=1.0.9
fake-useragent>=1.1.1
html5lib>=1.1
selenium>=4.9.0
webdriver_manager>=3.5.2
wikipedia-api>=0.5.4
wikipedia>=1.4.0
arxiv>=1.4.2
praw>=7.4.0  # For Reddit API
tweepy>=4.4.0  # For Twitter API

# Image Processing
Pillow>=8.3.2
opencv-python>=4.5.3
imagehash>=4.2.1

# Development
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.1
black>=21.9b0
flake8>=3.9.0
mypy>=0.910

# Optional dependencies for enhanced functionality
scikit-learn>=0.24.2  # For ML-based analysis
gensim>=4.1.2  # For topic modeling
fastapi>=0.68.1  # For API endpoints if needed
uvicorn>=0.15.0  # For serving API
python-multipart>=0.0.5  # For file uploads
python-jose[cryptography]>=3.3.0  # For JWT handling
passlib>=1.7.4  # For password hashing
bcrypt>=3.2.0  # For password hashing
  